[benc/item-editor-produces-item] Rename file to match class
