+++ packages/perseus-editor/src/__tests__/combined-hints-editor.test.tsx
@@ -2,7 +2,7 @@ import {render, screen} from "@testing-library/react";
 import {userEvent as userEventLib} from "@testing-library/user-event";
 import * as React from "react";
 
-import CombinedHintsEditor from "../hint-editor";
+import CombinedHintsEditor from "../combined-hints-editor";
 import {registerAllWidgetsAndEditorsForTesting} from "../util/register-all-widgets-and-editors-for-testing";
 
 import type {UserEvent} from "@testing-library/user-event";
